test(Coupon): add test cases
1. 移除src/coupon-list/test/index.legacy.js
2.
将原有测试用例分别对应到coupon、coupon-cell、coupon-list的test/
3. type
CouponInfo.id定义增加number类型
